fix(ci): python-coverage-shards — split residue and coverage-edge parts over the 1500-line gate
Review fixes grew comment_verdict_residue and parts 003/004 past MAX_FIRST_PARTY_FILE_LINES; extract nested metadata helpers and split oversized tests so the maintainability check passes again.
